refactor: change randomization service to RandomUtils
