+++ src/app/services/categories-service/categories.service.ts
@@ -2,7 +2,7 @@ import { inject, Injectable } from '@angular/core';
 import { environment } from '../../../environments/environment';
 import { catchError, finalize, map, Observable, tap } from 'rxjs';
 import { QuizCategory } from '../../shared/models/quiz-category.model';
-import { RandomizationService } from '../randomization-service/randomization.service';
+import { RandomUtils } from '../../utils/random';
 import { CategoriesStoreService } from '../categories-store-service/categories-store.service';
 import { ApiService } from '../api-service/api.service';
 import {
@@ -21,7 +21,6 @@ export class CategoriesService {
   private readonly maxNumberQuestions = environment.maxNumberOfQuestion;
 
   private readonly apiService = inject(ApiService);
-  private readonly randomizationService = inject(RandomizationService);
   private readonly categoriesStoreService = inject(CategoriesStoreService);
   private readonly errorHandlerService = inject(ErrorHandlerService);
 
@@ -29,15 +28,13 @@ export class CategoriesService {
     showSpinner();
     return this.apiService.fetchCategories().pipe(
       skipWhenCategoriesCached('quizCategories'),
-      map(categories => this.randomizationService
-      .getRandomItems(categories, this.categoriesNumber)
-      .map(category => this.enrichCategory(category))),
+      map(categories => RandomUtils.getRandomItems(categories, this.categoriesNumber).map(
+        category => this.enrichCategory(category),
+      )),
       tap((categories) => {
         this.categoriesStoreService.addCategories(categories);
       }),
-      catchError((error) => {
-        return this.errorHandlerService.handleError(error);
-      }),
+      catchError(this.errorHandlerService.handleError),
       finalize(() => {
         hideSpinner();
       }),
@@ -47,8 +44,8 @@ export class CategoriesService {
   private enrichCategory(category: QuizCategory): QuizCategory {
     return {
       ...category,
-      cardColor: this.randomizationService.getRandomColor(),
-      numberOfQuestion: this.randomizationService.getRandomInt(
+      cardColor: RandomUtils.getRandomColor(),
+      numberOfQuestion: RandomUtils.getRandomInt(
         this.minNumberQuestions,
         this.maxNumberQuestions,
       ),
